remove OnlyOneTypeOfKernel picker experiment
