+++ package.nls.json
@@ -372,10 +372,10 @@
     },
     "jupyter.configuration.jupyter.enableExtendedKernelCompletions.markdownDescription": "Enables Jedi support for extended IntelliSense completions in running Jupyter kernels (see this [setting](https://ipython.readthedocs.io/en/stable/config/options/terminal.html?highlight=use_jedi#configtrait-Completer.use_jedi)). This can greatly impact notebook cell execution performance. Use with caution.",
     "jupyter.configuration.jupyter.showOnlyOneTypeOfKernel.markdownDescription": "Show only one type of kernel in the kernel picker. This prevents local kernels and remote kernels from being shown at the same time in the kernel picker. Takes precedence over kernelPickerType if specified.",
+    "jupyter.configuration.jupyter.showOnlyOneTypeOfKernel.deprecationMessage": "This setting has been deprecated in favor of experimental.kernelPickerType.",
     "jupyter.configuration.jupyter.experimental.kernelPickerType.description": "Choose the type of kernel picker to use.",
-    "jupyter.configuration.jupyter.experimental.kernelPickerType.items.enumDescriptions.stable": "The current shipping kernel picker.",
-    "jupyter.configuration.jupyter.experimental.kernelPickerType.items.enumDescriptions.onlyonetypeofkernel": "Makes it so the picker will only show remote or local kernels, not both. And provides options to switch between local and remote kernels in the picker.",
-    "jupyter.configuration.jupyter.experimental.kernelPickerType.items.enumDescriptions.insiders": "Insiders is the newest experimental version in development.",
+    "jupyter.configuration.jupyter.experimental.kernelPickerType.items.enumDescriptions.stable": "The current stable kernel picker.",
+    "jupyter.configuration.jupyter.experimental.kernelPickerType.items.enumDescriptions.insiders": "Insiders is the new experimental version in development.",
     "Pylance.proposePylanceMessage": "Try out a new faster, feature-rich language server for Python by Microsoft, Pylance! Install the extension now.",
     "Pylance.tryItNow": "Try it now",
     "Pylance.remindMeLater": "Remind me later",
@@ -1118,9 +1118,6 @@
     "DataScience.outputSizeExceedLimit": "Output exceeds the <a href={0}>size limit</a>. Open the full output data <a href={1}>in a text editor</a>",
     "DataScience.installPythonTitle": "Install Python",
     "DataScience.installPythonExtensionViaKernelPickerTitle": "Install Python Extension",
-    "DataScience.switchToLocalKernelsTitle": "Connect to Local Kernels",
-    "DataScience.switchToRemoteKernelsTitle": "Connect to a Jupyter Server",
-    "DataScience.switchToAnotherRemoteKernelsTitle": "Connect to Another Jupyter Server",
     "DataScience.pickDocumentKernelSourceTitle": "Select Kernel Source",
     "DataScience.failedToInstallPythonExtension": "Failed to install the Python Extension.",
     "DataScience.kernelPrefixForRemote": "(Remote)",
