+++ src/commands.ts
@@ -202,8 +202,5 @@ export interface ICommandNameArgumentTypeMapping extends ICommandNameWithoutArgu
     [DSCommands.ReplayPylanceLogStep]: [];
     [DSCommands.InstallPythonExtensionViaKernelPicker]: [];
     [DSCommands.InstallPythonViaKernelPicker]: [];
-    [DSCommands.SwitchToLocalKernels]: [];
-    [DSCommands.SwitchToRemoteKernels]: [];
-    [DSCommands.SwitchToAnotherRemoteKernels]: [];
     [DSCommands.PickDocumentKernelSource]: [NotebookDocument | undefined];
 }
