+++ src/notebooks/controllers/kernelSource/notebookKernelSourceSelector.ts
@@ -15,7 +15,6 @@ import {
     IMultiStepInputFactory,
     IQuickPickParameters
 } from '../../../platform/common/utils/multiStepInput';
-import { compareIgnoreCase, groupBy } from '../commands/serverConnectionControllerCommands';
 import {
     IControllerRegistration,
     INotebookKernelSourceSelector,
@@ -164,4 +163,22 @@ export class NotebookKernelSourceSelector implements INotebookKernelSourceSelect
     toQuickPickItem(kernelFinderInfo: IContributedKernelFinderInfo): KernelFinderQuickPickItem {
         return { kernelFinderInfo, label: kernelFinderInfo.displayName };
     }
+}
+
+function groupBy<T>(data: ReadonlyArray<T>, compare: (a: T, b: T) => number): T[][] {
+    const result: T[][] = [];
+    let currentGroup: T[] | undefined = undefined;
+    for (const element of data.slice(0).sort(compare)) {
+        if (!currentGroup || compare(currentGroup[0], element) !== 0) {
+            currentGroup = [element];
+            result.push(currentGroup);
+        } else {
+            currentGroup.push(element);
+        }
+    }
+    return result;
+}
+
+function compareIgnoreCase(a: string, b: string) {
+    return a.localeCompare(b, undefined, { sensitivity: 'accent' });
 }
