+++ src/platform/common/configSettings.ts
@@ -231,13 +231,8 @@ export class JupyterSettings implements IWatchableJupyterSettings {
                   optOutFrom: []
               };
 
-        // For kernelPickerType internally collapse into just the kernelPickerType value
-        // Prefer the existing showOnlyOneTypeOfKernel value over the experimental picker
         const kernelPickerType = jupyterConfig.get<KernelPickerType>('experimental.kernelPickerType');
-        const showOnlyOneTypeOfKernel = jupyterConfig.get<boolean>('showOnlyOneTypeOfKernel');
-        if (showOnlyOneTypeOfKernel) {
-            this.kernelPickerType = 'OnlyOneTypeOfKernel';
-        } else if (kernelPickerType) {
+        if (kernelPickerType) {
             this.kernelPickerType = kernelPickerType;
         }
 
@@ -251,7 +246,7 @@ export class JupyterSettings implements IWatchableJupyterSettings {
             }
         };
         const keys = this.getSerializableKeys().filter(
-            (f) => f !== 'experiments' && f !== 'logging' && f !== 'kernelPickerType' && f !== 'showOnlyOneTypeOfKernel'
+            (f) => f !== 'experiments' && f !== 'logging' && f !== 'kernelPickerType'
         );
         keys.forEach((k) => replacer(k, jupyterConfig));
