+++ src/platform/common/constants.ts
@@ -284,9 +284,6 @@ export namespace Commands {
     export const ReplayPylanceLogStep = 'jupyter.replayPylanceLogStep';
     export const InstallPythonExtensionViaKernelPicker = 'jupyter.installPythonExtensionViaKernelPicker';
     export const InstallPythonViaKernelPicker = 'jupyter.installPythonViaKernelPicker';
-    export const SwitchToLocalKernels = 'jupyter.switchToLocalKernels';
-    export const SwitchToRemoteKernels = 'jupyter.switchToRemoteKernels';
-    export const SwitchToAnotherRemoteKernels = 'jupyter.switchToAnotherRemoteKernels';
     export const PickDocumentKernelSource = 'jupyter.pickDocumentKernelSource';
 }
