+++ src/test/datascience/notebook/helper.ts
@@ -39,7 +39,6 @@ import {
     QuickInputButton,
     QuickPickItemButtonEvent,
     EventEmitter,
-    ConfigurationTarget,
     NotebookEditor,
     debug,
     NotebookData
@@ -54,13 +53,7 @@ import {
 } from '../../../platform/common/constants';
 import { disposeAllDisposables } from '../../../platform/common/helpers';
 import { traceInfo, traceInfoIfCI, traceVerbose, traceWarning } from '../../../platform/logging';
-import {
-    GLOBAL_MEMENTO,
-    IConfigurationService,
-    IDisposable,
-    IMemento,
-    IsWebExtension
-} from '../../../platform/common/types';
+import { GLOBAL_MEMENTO, IDisposable, IMemento, IsWebExtension } from '../../../platform/common/types';
 import { createDeferred, sleep } from '../../../platform/common/utils/async';
 import {
     IKernelFinder,
@@ -1396,16 +1389,6 @@ export async function hijackPrompt(
         },
         clickButton: (text?: string) => clickButton.resolve(text || buttonToClick?.result)
     };
-}
-
-export async function changeShowOnlyOneTypeOfKernel(setting: boolean) {
-    const targetValue = setting ? 'OnlyOneTypeOfKernel' : 'Stable';
-    const api = await initialize();
-    const config = api.serviceContainer.get<IConfigurationService>(IConfigurationService);
-    const settings = config.getSettings();
-    if (settings.kernelPickerType !== targetValue) {
-        await config.updateSetting('experimental.kernelPickerType', targetValue, undefined, ConfigurationTarget.Global);
-    }
 }
 
 export async function hijackSavePrompt(
